Move entity framework methods
